Gazin Final Project_Pupil Tracking/requirements.txt
numpy==1.22.3
torch>=1.11
torchvision
Pillow
tqdm
opencv-contrib-python==4.5.5.64
matplotlib



